requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
idna==3.2
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
webdriver-manager==3.4.2
